refactor(runner): extract mapping to VS abstractions
