+++ sln/src/DotNetTestNSpec/DesignTime/DiscoveryRunner.cs
@@ -1,6 +1,4 @@
 ﻿using DotNetTestNSpec.Proxy;
-using Microsoft.Extensions.Testing.Abstractions;
-using System;
 
 namespace DotNetTestNSpec.DesignTime
 {
@@ -22,7 +20,7 @@ namespace DotNetTestNSpec.DesignTime
 
             foreach (var example in discoveredExamples)
             {
-                var test = MapToTest(example);
+                var test = MappingUtils.MapToTest(example);
 
                 adapter.TestFound(test);
             }
@@ -37,44 +35,5 @@ namespace DotNetTestNSpec.DesignTime
         readonly IDiscoveryAdapter adapter;
 
         const int dontCare = -1;
-
-        static Test MapToTest(DiscoveredExample example)
-        {
-            var test = new Test()
-            {
-                FullyQualifiedName = example.FullName,
-                DisplayName = BeautifyForDisplay(example.FullName),
-                CodeFilePath = example.SourceFilePath,
-                LineNumber = example.SourceLineNumber,
-            };
-
-            return test;
-        }
-
-        static string BeautifyForDisplay(string fullName)
-        {
-            // beautification idea taken from
-            // https://github.com/osoftware/NSpecTestAdapter/blob/master/NSpec.TestAdapter/TestCaseDTO.cs
-
-            string displayName;
-
-            // chop leading, redundant 'nspec. ' context, if any
-
-            const string nspecPrefix = @"nspec. ";
-            const int prefixLength = 7;
-
-            displayName = fullName.StartsWith(nspecPrefix, StringComparison.OrdinalIgnoreCase)
-                ? fullName.Substring(prefixLength)
-                : fullName;
-
-            // replace context separator
-
-            const string originalSeparator = @". ";
-            const string displaySeparator = @" › ";
-
-            displayName = displayName.Replace(originalSeparator, displaySeparator);
-
-            return displayName;
-        }
     }
 }
