+++ sln/src/DotNetTestNSpec/DesignTime/ExecutionRunner.cs
@@ -1,6 +1,5 @@
 ﻿using DotNetTestNSpec.Proxy;
 using Microsoft.Extensions.Testing.Abstractions;
-using System;
 using System.Collections.Generic;
 
 namespace DotNetTestNSpec.DesignTime
@@ -37,7 +36,7 @@ namespace DotNetTestNSpec.DesignTime
 
         public void ExampleStarted(DiscoveredExample example)
         {
-            var test = MapToTest(example);
+            var test = MappingUtils.MapToTest(example);
 
             startedTestMap[example.FullName] = test;
 
@@ -48,7 +47,7 @@ namespace DotNetTestNSpec.DesignTime
         {
             var test = startedTestMap[example.FullName];
 
-            var testResult = MapToTestResult(example, test);
+            var testResult = MappingUtils.MapToTestResult(example, test);
 
             adapter.TestFinished(testResult);
         }
@@ -60,69 +59,5 @@ namespace DotNetTestNSpec.DesignTime
         readonly IExecutionAdapter adapter;
 
         const int dontCare = -1;
-
-        static Test MapToTest(DiscoveredExample example)
-        {
-            var test = new Test()
-            {
-                FullyQualifiedName = example.FullName,
-                DisplayName = BeautifyForDisplay(example.FullName),
-                CodeFilePath = example.SourceFilePath,
-                LineNumber = example.SourceLineNumber,
-            };
-
-            return test;
-        }
-
-        static TestResult MapToTestResult(ExecutedExample example, Test test)
-        {
-            var testResult = new TestResult(test)
-            {
-                DisplayName = BeautifyForDisplay(example.FullName),
-                ErrorMessage = example.ExceptionMessage,
-                ErrorStackTrace = example.ExceptionStackTrace,
-            };
-
-            if (example.Pending)
-            {
-                testResult.Outcome = TestOutcome.Skipped;
-            }
-            else if (example.Failed)
-            {
-                testResult.Outcome = TestOutcome.Failed;
-            }
-            else
-            {
-                testResult.Outcome = TestOutcome.Passed;
-            }
-
-            return testResult;
-        }
-
-        static string BeautifyForDisplay(string fullName)
-        {
-            // beautification idea taken from
-            // https://github.com/osoftware/NSpecTestAdapter/blob/master/NSpec.TestAdapter/TestCaseDTO.cs
-
-            string displayName;
-
-            // chop leading, redundant 'nspec. ' context, if any
-
-            const string nspecPrefix = @"nspec. ";
-            const int prefixLength = 7;
-
-            displayName = fullName.StartsWith(nspecPrefix, StringComparison.OrdinalIgnoreCase)
-                ? fullName.Substring(prefixLength)
-                : fullName;
-
-            // replace context separator
-
-            const string originalSeparator = @". ";
-            const string displaySeparator = @" › ";
-
-            displayName = displayName.Replace(originalSeparator, displaySeparator);
-
-            return displayName;
-        }
     }
 }
